==> code/2 - ICD_Embeddings.ipynb <==
## Working with the csv file


import pandas as pd
pd.set_option('display.max_rows', 100)

from google.colab import drive
drive.mount('/content/drive')

%cd '/content/drive/MyDrive/softpr_sose23/Datasets'

df = pd.read_csv('/content/drive/MyDrive/softpr_sose23/Datasets/icd_defs_modified.csv')
df.head()

disease_names = []
for name in df["Disease"]:
  disease_names.append(name)

disease_names[-1]

## Pre-processing

Remove punctuations.

Convert all the texts to lowercase.

Tokenize the texts into individual words.

Remove stop words, such as "and", "the", "a", etc.

Perform lemmatization on the remaining words to convert them to their base form.

Filter out any words that occur infrequently in the corpus to reduce the dimensionality of the data.

Create word embeddings of the texts.

# pre-processing the descriptions for further work
import string
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
nltk.download("stopwords")
nltk.download('wordnet')

lemmatizer = WordNetLemmatizer()
stop_words = set(stopwords.words('english'))

df["Description_words"] = df["Description"].fillna('') + ' ' + df["Essential_Features"].fillna('')
df["Description_words"] = df["Description_words"].str.split()

df['Description_words'] = df['Description_words'].apply(lambda x: [lemmatizer.lemmatize(word.translate(str.maketrans('', '', string.punctuation)).lower()) for word in x if word.lower() not in stop_words])
df['Description_words'] = df['Description_words'].apply(lambda x: [word for word in x if not any(char.isdigit() for char in word)])
df['Description_words'] = df['Description_words'].apply(lambda words: ' '.join(words))

df['Description_words'][0]

# Obtaining the text embeddings

The process of creating word embeddings involves training a neural network on a large corpus of text data. However, pre-trained word embeddings are readily available online, and we utilize them in our project. See a complete list of pre-trained models [here](https://github.com/UKPLab/sentence-transformers/blob/master/docs/pretrained-models/sts-models.md).

!pip3 install tueplots==0.0.5
!pip3 install transformers
!pip install torch --upgrade
!pip3 install sentence-transformers==2.2.2

from sentence_transformers import SentenceTransformer

name = "stsb-bert-large"
model = SentenceTransformer(name)

sentences = list(df.Description_words.values)
disorder_embeddings = model.encode(sentences)

len(disorder_embeddings)

df['Embeddings'] = disorder_embeddings.tolist()

df.head()

df.to_csv('icd_embedded.csv', index=False)

# Projection of word embeddings on 3d space



disease_names = []
for name in df["Disease"]:
    disease_names.append(name)

import plotly.express as px
import pandas as pd
from sklearn.decomposition import PCA

# performing dimensionality reduction using PCA
pca = PCA(n_components=3)
embeddings_3d = pca.fit_transform(disorder_embeddings)

# creating a DataFrame for plotly
data_df = pd.DataFrame(embeddings_3d, columns=['PC1', 'PC2', 'PC3'])
data_df['Disease'] = disease_names

# creating an interactive 3D scatter plot
fig = px.scatter_3d(data_df, x='PC1', y='PC2', z='PC3', text='Disease')
fig.show()

# Projection of word embeddings on 2d space

# performing dimensionality reduction using PCA
pca = PCA(n_components=2)
embeddings_2d = pca.fit_transform(disorder_embeddings)

# creating a dataframe for plotly
data_df = pd.DataFrame(embeddings_2d, columns=['PC1', 'PC2'])
data_df['Disease'] = disease_names

# creating an interactive 2D scatter plot
fig = px.scatter(data_df, x='PC1', y='PC2', text='Disease')
fig.show()
